Remove unneeded MessageList.
