+++ packages/plugins/experimental/plugin-inbox/src/components/Mailbox/Mailbox.stories.tsx
@@ -1,0 +1,31 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import '@dxos-theme';
+import './mailbox.css';
+
+import { type Meta } from '@storybook/react';
+import React, { useState } from 'react';
+
+import { withLayout, withTheme } from '@dxos/storybook-utils';
+
+import { Mailbox } from './Mailbox';
+import { createInbox } from '../../testing';
+
+const DefaultStory = () => {
+  const [inbox] = useState(() => createInbox(100));
+
+  return <Mailbox mailbox={inbox} />;
+};
+
+export const Default = {};
+
+const meta: Meta = {
+  title: 'plugins/plugin-inbox/Mailbox',
+  component: Mailbox,
+  render: DefaultStory,
+  decorators: [withTheme, withLayout({ fullscreen: true })],
+};
+
+export default meta;
